requirements.txt
colorama
jinja2
requests
scapy
fpdf2
google